# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: NexusBridge HTTP API
  version: 0.1.0
  description: |
    NexusBridge 后端 HTTP API 契约。当前文件描述已实现接口，前端新需求先写入 docs/api/requests.md。
servers:
  - url: http://0.0.0.0:8090
    description: 本地默认服务
...
paths:
...
  /api/torrents:
    get:
      tags: [torrents]
      summary: 分页查询本地种子缓存
      parameters:
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: site_id
          in: query
          required: false
          schema:
            type: string
        - name: q
          in: query
          required: false
          schema:
            type: string
        - name: category
          in: query
          required: false
          description: 可重复的分类值；仅与单个 site_id 一起使用，同类值按 OR 匹配。
          style: form
          explode: true
          schema:
            type: array
            items:
              type: string
        - name: promotion
          in: query
          required: false
          description: 可重复的 spstate filter_value；仅与单个 site_id 一起使用，同类值按 OR 匹配。
          style: form
          explode: true
          schema:
            type: array
            items:
              type: string
        - name: site_checkbox
          in: query
          required: false
          description: 可重复的 group:value；仅与单个 site_id 一起使用，同组值按 OR、不同组按 AND 匹配。
          style: form
          explode: true
          schema:
            type: array
            items:
              type: string
        - name: qb_task
          in: query
          required: false
          description: 按本地种子是否关联当前 qB 任务筛选；列表请求不会主动访问 qB。
          schema:
            type: string
            enum: [present, absent]
        - name: qb_progress
          in: query
          required: false
          description: 仅与 qb_task=present 一起使用；complete 为 progress >= 1。
          schema:
            type: string
            enum: [complete, incomplete]
        - name: include_pinned
          in: query
          required: false
          schema:
            type: boolean
            default: true
        - name: sort_by
          in: query
          required: false
          schema:
            type: string
            enum: [source_order, published_at, size_bytes, seeders, leechers, snatches]
            default: published_at
        - name: sort_direction
          in: query
          required: false
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        "200":
          description: 筛选、排序和范围截取后的种子页
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TorrentPage"
        "400":
          $ref: "#/components/responses/Error"
        "409":
          description: progress 筛选所需的 qB 运行态尚未完成首次同步
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/Error"
  /api/playback/torrents:
    get:
      tags: [torrents, qbittorrent]
      summary: 查询含有完整音频或视频文件的其他 qB 种子
      parameters:
        - name: exclude_site_id
          in: query
          required: false
          schema:
            type: string
        - name: exclude_torrent_id
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
      responses:
        "200":
          description: 按站点发布时间倒序的其他可播放种子
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PlaybackTorrent"
        "400":
          $ref: "#/components/responses/Error"
        "503":
          $ref: "#/components/responses/Error"
...
  /api/qb/torrents/{hash}:
    parameters:
      - name: hash
        in: path
        required: true
        description: 单个 40 或 64 位十六进制 info hash；不接受 all。
        schema:
          type: string
          pattern: "^[0-9A-Fa-f]{40}([0-9A-Fa-f]{24})?$"
    delete:
      tags: [qbittorrent]
      summary: 删除单个 qB 任务并选择是否同时删除下载文件
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QBTorrentDeleteRequest"
      responses:
        "200":
          description: qB 任务已删除
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "502":
          $ref: "#/components/responses/Error"
...
components:
...
  responses:
    Error:
      description: 错误响应
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
    DeleteResult:
      type: object
      required: [deleted]
      properties:
        deleted:
          type: boolean
...
    Torrent:
      type: object
      required: [id, site_id, title, detail_url, first_seen_at, last_seen_at, source_order, sticky_level, torrent_file_saved]
      properties:
        id:
          type: string
        site_id:
          type: string
        category:
          type: string
        category_query:
          type: string
        title:
          type: string
        detail_url:
          type: string
        download_url:
          type: string
        cover_url:
          type: string
        tags:
          type: array
          items:
            type: string
        tag_ids:
          type: array
          items:
            type: string
        description:
          type: string
        detail_title:
          type: string
        subtitle:
          type: string
        product_url:
          type: string
        detail_info_hash:
          type: string
        detail_description:
          type: string
        detail_raw_text:
          type: string
        detail_fetched_at:
          type: string
        promotion:
          type: string
        promotion_class:
          type: string
        promotion_ends_at:
          type: string
        size_bytes:
          type: integer
          format: int64
        seeders:
          type: integer
        leechers:
          type: integer
        snatches:
          type: integer
        comments:
          type: integer
        published_at:
          type: string
          format: date-time
        published_text:
          type: string
        first_seen_at:
          type: string
          format: date-time
        last_seen_at:
          type: string
          format: date-time
        source_order:
          type: integer
          description: 最近一次站点列表抓取中的从上到下序号
        sticky_level:
          type: integer
          minimum: 0
          description: 当前进程最近一次站点抓取的内存置顶等级；大于 0 时数值较高者优先。
        torrent_file_saved:
          type: boolean
        info_hash_v1:
          type: string
        info_hash_v2:
          type: string
        torrent_file_error:
          type: string
        qb_status:
          $ref: "#/components/schemas/QBTorrentStatus"
    TorrentPage:
      type: object
      required: [items, offset, limit, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Torrent"
        offset:
          type: integer
          minimum: 0
        limit:
          type: integer
          minimum: 1
          maximum: 100
        total:
          type: integer
          minimum: 0
...
    PlaybackTorrent:
      type: object
      required: [id, site_id, title]
      properties:
        id:
          type: string
        site_id:
          type: string
        title:
          type: string
        cover_url:
          type: string
        category:
          type: string
        published_at:
          type: string
          format: date-time
    QBTorrentStatus:
      type: object
      required: [available, added, source, fetched_at]
      properties:
        available: { type: boolean }
        added: { type: boolean }
        source: { type: string }
        error: { type: string }
        fetched_at: { type: string, format: date-time }
        task_id: { type: string }
        task_status: { type: string }
        rule_name: { type: string }
        hash: { type: string }
        name: { type: string }
        state: { type: string }
        progress: { type: number, format: double }
        category: { type: string }
        tags: { type: string }
        save_path:
          type: string
          description: qB 返回的任务保存路径，可能是本机绝对路径。
        content_path:
          type: string
          description: qB 返回的任务内容路径，可能是本机绝对路径。
        download_speed: { type: integer, format: int64 }
        upload_speed: { type: integer, format: int64 }
        eta: { type: integer, format: int64 }
        ratio: { type: number, format: double }
        size: { type: integer, format: int64 }
        completed: { type: integer, format: int64 }
        amount_left: { type: integer, format: int64 }
...
    QBTorrentDeleteRequest:
      type: object
      required: [delete_files]
      properties:
        delete_files:
          type: boolean
          description: true 时由 qBittorrent 同时删除已下载文件；false 时保留文件。
...